Introduce new `datadog_ruby_common.h/.c` for common bits between extensions
This cleans up the wad of copy-pasted code I did for my first attempt.
